Win32-utf8 * reduce some warning while compiling * add static keyword to functions in ChewingServer
* char.c dict.c tree.c
use open (POSIX API) to replace CreateHandle(Win32 API)
It become the same as trunk, but Win32 don't support mmap.

* Apply wiz's patch: src/zuin.c (PinYinInput):
It should clean the buffer of PinYin data and return "ZUIN_ABSORB"
to ignore the current key when converting failed.



git-svn-id: https://svn.csie.net/chewing/win32-chewing@656 ac1be623-90ea-0310-9410-ba68b2efa777

+++ ChewingIME/CandList.cpp
@@ -4,7 +4,7 @@ CandList::CandList(void)
 {
 	ci.dwSize = sizeof(CandList);
 	ci.dwCount = 1;
-	ci.dwOffset[0] = (size_t(&cl) - size_t(&ci));
+	ci.dwOffset[0] = (DWORD) (size_t(&cl) - size_t(&ci));
 	ci.dwPrivateSize = 0;
 
 	cl.dwSize = sizeof(cl)+sizeof(offset)+sizeof(candStr);
